Add memory leak diagnostic
I noticed that in some cases, the stream fails to start and when I SSH
back in, I see that the application is failing with MemoryErrors. To
diagnose this, add a regular timer which fetches some memory values for
the application and system, and overlays it onto the video. Also, it
displays the number of connections. This should allow me to monitor the
application and see what scenarios trigger the memory leak.