Fix memory leak
When a client disconnects, normally the browser would politely tear down
the websocket connection and Tornado would detect this, closing the
connection. However, if a client stops responding without warning, it
seems that Tornado does not notice this, and it just keeps buffering
messages for the client.

I've noticed this in two cases. First, a laptop falling asleep may
trigger unreasonable server-side buffering. Second, my phone
disconnecting from the VPN will do the same.

To address this, set the Tornado websocket_ping_{timeout,interval}
application settings. This causes Tornado to send ping websocket
protocol-level ping messages, which the browser automatically replies
to. If the browser hasn't replied in the last <interval> seconds, then
the connection is torn down. This ensures that misbehaving clients can't
trigger the memory leak.

I set these settings to values which are appropriate to my use case, but
they should probably be easier to configure.