Add settings endpoint
This allows us to dynamically configure camera settings (anything which
can be configured at runtime). This is just the backend, it can be
curl'd with ease. The frontend will be a form with a submit button and
some zoom buttons.

The goal here is to make it easy to compensate for poor lighting, and
also compensate for the zoomed-out fisheye lens I have.