Add client interface for the camera settings
I'm no frontend dev (and this commit probably shows it).

Previously I added the /cam/ API to get/set settings and zoom
information. This adds forms & buttons to make it work from the client.
We have two major pieces: first, a form inside a pop-up modal, for
editing the different levels (brightness, contrast, etc) along with the
modes (white balance, image effect). The second piece is a button matrix
that can zoom in/out or pan the zoomed image. One of the buttons can
call up the settings modal.

Since I'm no frontend dev, I've included the farm: Bootstrap, jQuery,
Font Awesome, and the little bootstrap JS libraries as well. Bootstrap
makes the button styles nice and makes the modals easy. I'm a bit scared
of doing CSS without Bootstrap, to be honest.

Anyhow, the result is not terrible. It works and it's interesting for
exploring the camera settings (the ones which work, anyway).